+++ src/app/components/common/header/menu.tsx
@@ -3,8 +3,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { useEffect } from 'react';
-import { useRouter } from 'next-nprogress-bar';
-import { usePathname, useSearchParams } from 'next/navigation';
+import { usePathname } from 'next/navigation';
 import { useShallow } from 'zustand/react/shallow';
 import classNames from 'classnames';
 import { AnimatePresence, motion } from 'framer-motion';
@@ -12,45 +11,18 @@ import { IoCloseOutline } from 'react-icons/io5';
 import { HiEllipsisVertical } from 'react-icons/hi2';
 import { GrFormNextLink } from 'react-icons/gr';
 
+import MenuFilter from './menu-filter';
 import useTableStore from '@/app/stores/table';
-import usePlayerStore from '@/app/stores/player';
-import useBuffStore from '@/app/stores/buff';
-import useCommonStore from '@/app/stores/common';
 
 export default function Menu() {
   const [isMenu, openMenu, closeMenu] = useTableStore(
     useShallow((state) => [state.isMenu, state.openMenu, state.closeMenu])
   );
-  const [setSelectedPlayer, setPinnedPlayer, clearLineup] = usePlayerStore(
-    useShallow((state) => [state.setSelectedPlayer, state.setPinnedPlayer, state.clearLineup])
-  );
-  const clearBuff = useBuffStore((state) => state.clearBuff);
-  const setIsLoading = useCommonStore((state) => state.setIsLoading);
-  const router = useRouter();
   const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const buttonValueList = [55, 60, 65, 69];
-  const limit = searchParams.get('limit');
-  const overallLimit = !limit || isNaN(+limit) || +limit > 99 ? 69 : +limit < 55 ? 55 : +limit;
 
   useEffect(() => {
     closeMenu();
   }, [pathname]);
-
-  useEffect(() => {
-    setSelectedPlayer(null);
-    setPinnedPlayer(null);
-    clearLineup();
-    clearBuff();
-    setIsLoading(false);
-  }, [overallLimit]);
-
-  const onFilterClick = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    if (overallLimit === Number(e.currentTarget.value)) return;
-
-    setIsLoading(true);
-    router.replace(`${pathname}?limit=${Number(e.currentTarget.value)}`, { scroll: false });
-  };
 
   return (
     <>
@@ -82,38 +54,7 @@ export default function Menu() {
               </button>
             </div>
 
-            <div
-              className={classNames('mb-20 border-b-1 border-slate-200/10 pb-20', {
-                hidden: !pathname.startsWith('/entry/'),
-              })}
-            >
-              <span className='mb-10 block text-18 font-semibold'>오버롤 설정</span>
-
-              <span className='mb-10 block text-12 text-red-300 opacity-80'>
-                오버롤 필터 변경 시 선택한 선수들이 초기화됩니다.
-              </span>
-
-              <ul className='flex flex-wrap items-center justify-center gap-5'>
-                {buttonValueList.map((value) => (
-                  <li key={value} className='flex-1'>
-                    <button
-                      className={classNames(
-                        'aspect-square w-full rounded-md border-1 text-15 transition-[color,_background-color,_opacity] duration-200',
-                        {
-                          'bg-slate-100 font-semibold text-[#e56d36]': value === overallLimit,
-                          'text-slate-400 opacity-50 hover:bg-slate-300 hover:text-slate-800 hover:opacity-90':
-                            value !== overallLimit,
-                        }
-                      )}
-                      value={value}
-                      onClick={onFilterClick}
-                    >
-                      {value === 55 ? '전체' : `${value} 이상`}
-                    </button>
-                  </li>
-                ))}
-              </ul>
-            </div>
+            <MenuFilter />
 
             <div className='mb-20 border-b-1 border-slate-200/10 pb-20'>
               <Link href='/news'>
